Refactor search controller to move api client code into separate client file
